Move travel components to own folder
